# Refactor for faster assembly of secondHashTable.