--- a/src/mlpack/core/data/split_data.hpp
+++ b/src/mlpack/core/data/split_data.hpp
@@ -1,18 +1,17 @@
 /**
  * @file split_data.hpp
- * @author Tham Ngap Wei
+ * @author Tham Ngap Wei, Keon Kim
  *
- * Defines TrainTestSplit(), a utility function to split a dataset into a
+ * Defines Split(), a utility function to split a dataset into a
  * training set and a test set.
  */
-#ifndef MLPACK_CORE_UTIL_SPLIT_DATA_HPP
-#define MLPACK_CORE_UTIL_SPLIT_DATA_HPP
+#ifndef MLPACK_CORE_DATA_SPLIT_DATA_HPP
+#define MLPACK_CORE_DATA_SPLIT_DATA_HPP
 
 #include <mlpack/core.hpp>
 
 namespace mlpack {
 namespace data {
-
 /**
  * Given an input dataset and labels, split into a training set and test set.
  * Example usage below.  This overload places the split dataset into the four
@@ -29,7 +28,7 @@ namespace data {
  *
  * // Split the dataset into a training and test set, with 30% of the data being
  * // held out for the test set.
- * TrainTestSplit(input, label, trainData,
+ * Split(input, label, trainData,
  *                testData, trainLabel, testLabel, 0.3);
  * @endcode
  *
@@ -42,13 +41,13 @@ namespace data {
  * @param testRatio Percentage of dataset to use for test set (between 0 and 1).
  */
 template<typename T, typename U>
-void TrainTestSplit(const arma::Mat<T>& input,
-                    const arma::Row<U>& inputLabel,
-                    arma::Mat<T>& trainData,
-                    arma::Mat<T>& testData,
-                    arma::Row<U>& trainLabel,
-                    arma::Row<U>& testLabel,
-                    const double testRatio)
+void Split(const arma::Mat<T>& input,
+           const arma::Row<U>& inputLabel,
+           arma::Mat<T>& trainData,
+           arma::Mat<T>& testData,
+           arma::Row<U>& trainLabel,
+           arma::Row<U>& testLabel,
+           const double testRatio)
 {
   const size_t testSize = static_cast<size_t>(input.n_cols * testRatio);
   const size_t trainSize = input.n_cols - testSize;
@@ -75,6 +74,52 @@ void TrainTestSplit(const arma::Mat<T>& input,
 }
 
 /**
+ * Given an input dataset, split into a training set and test set.
+ * Example usage below. This overload places the split dataset into the two
+ * output parameters given (trainData, testData).
+ *
+ * @code
+ * arma::mat input = loadData();
+ * arma::mat trainData;
+ * arma::mat testData;
+ * math::RandomSeed(100); // Set the seed if you like.
+ *
+ * // Split the dataset into a training and test set, with 30% of the data being
+ * // held out for the test set.
+ * Split(input, trainData, testData, 0.3);
+ * @endcode
+ *
+ * @param input Input dataset to split.
+ * @param trainData Matrix to store training data into.
+ * @param testData Matrix to store test data into.
+ * @param testRatio Percentage of dataset to use for test set (between 0 and 1).
+ */
+template<typename T>
+void Split(const arma::Mat<T>& input,
+           arma::Mat<T>& trainData,
+           arma::Mat<T>& testData,
+           const double testRatio)
+{
+  const size_t testSize = static_cast<size_t>(input.n_cols * testRatio);
+  const size_t trainSize = input.n_cols - testSize;
+  trainData.set_size(input.n_rows, trainSize);
+  testData.set_size(input.n_rows, testSize);
+
+  const arma::Col<size_t> order =
+      arma::shuffle(arma::linspace<arma::Col<size_t>>(0, input.n_cols -1,
+                                                      input.n_cols));
+
+  for (size_t i = 0; i != trainSize; ++i)
+  {
+    trainData.col(i) = input.col(order[i]);
+  }
+  for (size_t i = 0; i != testSize; ++i)
+  {
+    testData.col(i) = input.col(order[i + trainSize]);
+  }
+}
+
+/**
  * Given an input dataset and labels, split into a training set and test set.
  * Example usage below.  This overload returns the split dataset as a std::tuple
  * with four elements: an arma::Mat<T> containing the training data, an
@@ -84,37 +129,62 @@ void TrainTestSplit(const arma::Mat<T>& input,
  * @code
  * arma::mat input = loadData();
  * arma::Row<size_t> label = loadLabel();
- * auto splitResult = TrainTestSplit(input, label, 0.2);
+ * auto splitResult = Split(input, label, 0.2);
  * @endcode
  *
  * @param input Input dataset to split.
  * @param label Input labels to split.
- * @param trainData Matrix to store training data into.
- * @param testData Matrix to store test data into.
- * @param trainLabel Vector to store training labels into.
- * @param testLabel Vector to store test labels into.
  * @param testRatio Percentage of dataset to use for test set (between 0 and 1).
  * @return std::tuple containing trainData (arma::Mat<T>), testData
  *      (arma::Mat<T>), trainLabel (arma::Row<U>), and testLabel (arma::Row<U>).
  */
 template<typename T,typename U>
 std::tuple<arma::Mat<T>, arma::Mat<T>, arma::Row<U>, arma::Row<U>>
-TrainTestSplit(const arma::Mat<T>& input,
-               const arma::Row<U>& inputLabel,
-               const double testRatio)
+Split(const arma::Mat<T>& input,
+      const arma::Row<U>& inputLabel,
+      const double testRatio)
 {
   arma::Mat<T> trainData;
   arma::Mat<T> testData;
   arma::Row<U> trainLabel;
   arma::Row<U> testLabel;
 
-  TrainTestSplit(input, inputLabel, trainData, testData, trainLabel, testLabel,
+  Split(input, inputLabel, trainData, testData, trainLabel, testLabel,
       testRatio);
 
   return std::make_tuple(std::move(trainData),
                          std::move(testData),
                          std::move(trainLabel),
                          std::move(testLabel));
+}
+
+/**
+ * Given an input dataset, split into a training set and test set.
+ * Example usage below.  This overload returns the split dataset as a std::tuple
+ * with two elements: an arma::Mat<T> containing the training data and an
+ * arma::Mat<T> containing the test data.
+ *
+ * @code
+ * arma::mat input = loadData();
+ * auto splitResult = Split(input, 0.2);
+ * @endcode
+ *
+ * @param input Input dataset to split.
+ * @param testRatio Percentage of dataset to use for test set (between 0 and 1).
+ * @return std::tuple containing trainData (arma::Mat<T>)
+ *      and testData (arma::Mat<T>).
+ */
+template<typename T>
+std::tuple<arma::Mat<T>, arma::Mat<T>>
+Split(const arma::Mat<T>& input,
+      const double testRatio)
+{
+  arma::Mat<T> trainData;
+  arma::Mat<T> testData;
+  Split(input, trainData, testData, testRatio);
+
+  return std::make_tuple(std::move(trainData),
+                         std::move(testData));
 }
 
 } // namespace data