--- a/src/mlpack/core/dists/gaussian_distribution.cpp
+++ b/src/mlpack/core/dists/gaussian_distribution.cpp
@@ -6,6 +6,7 @@
  * Implementation of Gaussian distribution class.
  */
 #include "gaussian_distribution.hpp"
+#include <mlpack/methods/gmm/positive_definite_constraint.hpp>
 
 using namespace mlpack;
 using namespace mlpack::distribution;
@@ -116,18 +117,7 @@ void GaussianDistribution::Train(const arma::mat& observations)
   covariance /= (observations.n_cols - 1);
 
   // Ensure that the covariance is positive definite.
-  if (det(covariance) <= 1e-50)
-  {
-    Log::Debug << "GaussianDistribution::Train(): Covariance matrix is not "
-        << "positive definite. Adding perturbation." << std::endl;
-
-    double perturbation = 1e-30;
-    while (det(covariance) <= 1e-50)
-    {
-      covariance.diag() += perturbation;
-      perturbation *= 10; // Slow, but we don't want to add too much.
-    }
-  }
+  gmm::PositiveDefiniteConstraint::ApplyConstraint(covariance);
 
   FactorCovariance();
 }
@@ -173,7 +163,8 @@ void GaussianDistribution::Train(const arma::mat& observations,
   }
 
   // Normalize.
-  mean /= sumProb;
+  if (sumProb > 0)
+    mean /= sumProb;
 
   // Now find the covariance.
   for (size_t i = 0; i < observations.n_cols; i++)
@@ -183,21 +174,11 @@ void GaussianDistribution::Train(const arma::mat& observations,
   }
 
   // This is probably biased, but I don't know how to unbias it.
-  covariance /= sumProb;
+  if (sumProb > 0)
+    covariance /= sumProb;
 
   // Ensure that the covariance is positive definite.
-  if (det(covariance) <= 1e-50)
-  {
-    Log::Debug << "GaussianDistribution::Train(): Covariance matrix is not "
-        << "positive definite. Adding perturbation." << std::endl;
-
-    double perturbation = 1e-30;
-    while (det(covariance) <= 1e-50)
-    {
-      covariance.diag() += perturbation;
-      perturbation *= 10; // Slow, but we don't want to add too much.
-    }
-  }
+  gmm::PositiveDefiniteConstraint::ApplyConstraint(covariance);
 
   FactorCovariance();
 }