@mentekid: a first attempt, not 100% tested. Let me know what you think. I should have time tonight to do further testing and comparison but it appears to be good so far.