@rcurtin 2016-06-04
I had fun rewriting it too, just ran out of time for now to test it. :)