@mentekid 2016-06-04
Here's what I have in mind regarding an std::vector<size_t> array.

This requires more work, but let me know what you think.

If you run this with default parameters it will probably take a lot of
time, that's because the old implementation discards a lot of points while
this way we don't discard any.

We could also implement an optional size limit for the vectors for backward
compatibility, for example set the default to actually be 500 points but if
you set bucketSize to 0 the vectors grow arbitrarily large.

On Sat, Jun 4, 2016 at 3:43 AM, Ryan Curtin <notifications@github.com>
wrote:

> I had fun rewriting it too, just ran out of time for now to test it. :)
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/mlpack/mlpack/pull/675#issuecomment-223725394>, or mute
> the thread
> <https://github.com/notifications/unsubscribe/AFxX3Co0w7vNpLCoCljOORwjhzLR2yQvks5qIMopgaJpZM4IuAcT>
> .
>