@rcurtin 2016-06-09
Okay, this is ready to go.  It'll need to be merged manually but that is no problem...

Here are some benchmarks as measured by mlpack_lsh on both the old and the new version.  The results are significant enough I didn't bother with more than one trial or tuning parameters.

corel.csv: hash building: 0.474s -> 0.293s, computing neighbors 28.057s -> 5.832s
isolet.csv: hash building 0.614s -> 0.517s, computing neighbors 0.776s -> 0.565s
covertype.csv: hash building 5.173s -> 6.215s, computing neighbors 1739.332s -> 1172.641s
phy.csv: hash building 2.074s -> 1.823s, computing neighbors 167.203s -> 51.062s
pokerhand.csv: hash building 0.329s -> 0.212s, computing neighbors 1.101s -> 0.737s
miniboone.csv: hash building 1.311s -> 1.429s, computing neighbors 51.144s -> 25.309s